frontend/pyproject.toml
[project]
name = "streamlit-page-navigation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit==1.50.0",
    "msal==1.34.0",
    "python-dotenv>=1.0.0",
    "requests>=2.25.0",
    "st_mui_table>=0.0.6",
    "pre-commit>=4.3.0",
    "ruff>=0.14.2",
    "python-docx>=1.2.0",
    "fpdf2>=2.8.4",
    "openpyxl>=3.1.5",
    "pip==25.3",
    "black>=25.11.0",
    "streamlit-pdf-viewer>=0.0.26",
    "st-copy>=0.1.0",
    "opentelemetry-api>=1.10.0",
    "opentelemetry-sdk>=1.10.0",
    "azure-monitor-opentelemetry",
    "azure-monitor-opentelemetry-exporter",
    "opencensus>=0.11.2",
    "opencensus-ext-azure>=1.0.0",
    "azure-identity-broker>=1.3.0",
]
